fix: theme switcher toggling incorrectly (#553)
Currently, when the theme is set to be either light or dark, and the browser's option is changed, it results in the theme being inverted with what is wanted (i.e. setting light theme results in dark theme, and vice versa).
With this change, the theme switcher works as expected.